requirements.txt
pytorch
torchvision
scikit-learn
pandas
numpy
py-opencv 
opencv
pillow
tensorboard
matplotlib
medpy
tqdm
segmentation-models-pytorch
timm